+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changes
 - Fix nv_base_tgp being referenced as dgpu_base_tgp
+- Add nv_tgp control slider
 
 ## [6.3.4]